Replace set for listing keys with `KEYS`
Setting a lot of keys with a TTL quickly fills up the `namespace:keyv`
with key names that no longer exist. If these keys never get read again,
they will stay stored in redis indefinitely.

This replaces the `namespace:keyv` set with a call to `KEYS keyv:*` to
get a list of keys that are currently set.